Update Cubefield.xml
Fixed spawns (I never told you I wanted them to be up higher, not your fault), changed the objective to be more descriptive, and added a tip. I think the XML is missing the filter that makes it so you can only enter the scorebox when you have control of the hill though.